requirements.txt
reportlab==3.6.1
python-barcode==0.13.1
